requirements.txt
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
openai==1.3.5
tweepy==4.14.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.3
pyyaml==6.0.1
jinja2==3.1.2
click==8.1.7
